requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
gradio>=3.35.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
accelerate>=0.20.0
evaluate>=0.4.0

# Optional: TensorBoard for logging
tensorboard>=2.13.0

# Development (optional)
jupyter>=1.0.0
ipykernel>=6.25.0 